more cleanup + simplify layout
