+++ lib/field-collection.js
