+++ lib/type-collection.js
