+++ lib/type.js
