+++ lib/value-collection.js
