+++ lib/value.js
